@@ -1,78 +1,48 @@
-import { Injectable, UnauthorizedException, HttpStatus } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { LoginUserDto } from '../users/dto/login-user.dto';
 import { UsersService } from '../users/users.service';
-import { JwtPayload } from './jwt-payload.interface';
+import { JwtPayload } from './interfaces/jwt-payload.interface';
 
 @Injectable()
 export class AuthService {
 
-    constructor(private usersService: UsersService, private jwtService: JwtService){
+    constructor(private usersService: UsersService, private jwtService: JwtService){ }
 
-    }
+    async validateUserByPassword(loginAttempt: LoginUserDto): Promise<any> {
+        let userToAttempt: any = await this.usersService.findOneByEmail(loginAttempt.email);
 
-    async validateUserByPassword(loginAttempt: LoginUserDto) {
-
-        // This will be used for the initial login
-        let userToAttempt = await this.usersService.findOneByEmail(loginAttempt.email);
-        
         return new Promise((resolve) => {
-
-            // Check the supplied password against the hash stored for this email address
+            if (!userToAttempt) {
+                resolve({success: false, msg: 'User not found'});
+            }
             userToAttempt.checkPassword(loginAttempt.password, (err, isMatch) => {
-    
-                if(err)      throw new UnauthorizedException();
+                if(err) resolve({success: false, msg: 'Unexpected error. Please try again later.'});
     
                 if(isMatch){
-                    // If there is a successful match, generate a JWT for the user
-                    resolve(this.createJwtPayload(userToAttempt));
-    
+                    resolve({success: true, data: this.createJwtPayload(userToAttempt)});
                 } else {
-                    throw new UnauthorizedException();
+                    resolve({success: false, msg: 'Wrong password'})
                 }
-    
             });
-
         });
-
     }
 
-    async validateUserByJwt(payload: JwtPayload) { 
-
-        // This will be used when the user has already logged in and has a JWT
-        let user = await this.usersService.findOneByEmail(payload.email);
-
-        if(user){
-            return this.createJwtPayload(user);
-        } else {
-            throw new UnauthorizedException();
-        }
-
-    }
-
-    createJwtPayload(user: JwtPayload){
-
+    createJwtPayload(user){
         let data: JwtPayload = {
-            email: user.email,
-            dogOwner: user.dogOwner,
-            _id: user._id,
-            dogName: user.dogName,
-            dogBreed: user.dogBreed,
-            userTodos: user.userTodos
+            id: user._id,
+            email: user.email
         };
 
         let jwt = this.jwtService.sign(data);
 
         return {
-            expiresIn: 60,
-            dogOwner: user.dogOwner,
-            token: jwt,
-            _id: user._id,
-            dogName: user.dogName,
-            dogBreed: user.dogBreed,
-            userTodos: user.userTodos
+            exp: 36000,
+            token: jwt            
         }
-
     }
 
+    async validateUser(payload: JwtPayload): Promise<any> {
+        return await this.usersService.getUser(payload.id);
+    }
 }
