@@ -1,17 +1,17 @@
 import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-import { PassportModule } from '@nestjs/passport';
-import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-import { UserSchema } from './user.schema';
+import { UsersController } from './users.controller';
+import { MongooseModule } from '@nestjs/mongoose';
+import { UserSchema } from './schemas/user.schema';
+import { PassportModule } from '@nestjs/passport';
 
 @Module({
   imports: [
-    MongooseModule.forFeature([{name: 'User', schema: UserSchema}]),
+    MongooseModule.forFeature([{ name: 'User', schema: UserSchema}]),
     PassportModule.register({ defaultStrategy: 'jwt', session: false })
   ],
-  exports: [UsersService],
+  providers: [UsersService],
   controllers: [UsersController],
-  providers: [UsersService]
+  exports: [UsersService]
 })
 export class UsersModule {}
