@@ -8,6 +8,7 @@ import { Router } from "@angular/router";
 
 import { SecureStorage, SecureStorageObject } from '@ionic-native/secure-storage/ngx';
 
+import { ApiService, User } from '../../services/api.service';
 
 @Component({
   selector: 'app-home',
@@ -19,49 +20,85 @@ export class HomePage {
   dogName:string = "/";
   dogBreed:string = "/";
   todos:Array<string> = [];
+  user:User;
 
-  constructor(private alert:AlertController, private httpReq: HttpRequestService, private modalController: ModalController, private router:Router, private secureStorage: SecureStorage){
-    this.todos = this.httpReq.getTodos();
+  constructor(private alert:AlertController, private api:ApiService, private httpReq: HttpRequestService, private modalController: ModalController, private router:Router, private secureStorage: SecureStorage){
   }
 
   ngOnInit(){
     console.log("hello from home page");
 
-    this.dogName = this.httpReq.getDogData().dogName;
-    this.dogBreed = this.httpReq.getDogData().dogBreed;
+    this.loadUserData();
+
     
+     
   }
 
-  async openTOdo(){
+  loadUserData(){
+    this.api.getUserData().subscribe(a =>{
+      this.user = a;
 
-      const alert = await this.alert.create({
-        header: `Please add task`,
-        inputs:[
-          {
-            name: "todo",
-            type: "text"
-          }
-        ],
-          buttons: [
-            {
-              text: 'Ok',
-              handler: (a) => {
-                this.todos.push(a.todo);
-                this.httpReq.saveUserTodo(this.todos);
+      if(a.dogBreed == "" && a.dogName == "" || a.dogBreed == null && a.dogName == null){
+        this.openAddDog();
+      }else{
+        this.dogBreed = this.user.dogBreed;
+        this.dogName = this.user.dogName;
+        this.todos = this.user.todos;
+      }
+
+    })
+  }
+  
+  async openTOdo(){
+    
+    const alert = await this.alert.create({
+      header: `Please add task`,
+      inputs:[
+        {
+          name: "todo",
+          type: "text"
+        }
+      ],
+      buttons: [
+        {
+          text: 'Ok',
+          handler: (a) => {
+            this.todos.push(a.todo);
+                this.api.updateUser(this.user._id, {"todos": this.todos}).subscribe(a => {
+                  console.log(a);
+                })
               }
             }
           ]})
           await alert.present();
-  }
-
+        }
+        
   logOut(){
-    this.router.navigate(['login']);
+    this.router.navigate(['']);
   }
 
+  signOut() {
+    this.api.logout();
+  }
+  
   delete(i:number){
     this.todos.splice(i, 1);
   }
 
+  async openAddDog(){
+    const alert = await this.alert.create({
+      header: `Would you like to add a dog?`,
+        buttons: [
+          {
+            text: 'Yes',
+            handler: () => {
+              this.router.navigateByUrl('/dog-form');
+            }
+          },
+          {text: "No"}
+        ]})
+        await alert.present();
+  }
   
-
+  
 }
